requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Mail
python-dotenv
requests
beautifulsoup4
APScheduler
email-validator
pytz
WTForms
groq